% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{alook_dependency}
\alias{alook_dependency}
\title{Functions}
\usage{
alook_dependency()
}
\description{
\code{apath} dependency.
}
\examples{
library(aframer)

a_scene(
  a_dependency(),
  alook_dependency()
)

}
